Phase 6b: Extract team data functions into useTeamData hook
Extract ~1,100 lines of team data fetching logic from App.jsx into a
dedicated useTeamData hook and a calculateBlueBanners utility module.

New files:
- src/hooks/useTeamData.js (1,235 lines): Contains getTeamList,
  getEPA, getEPAFTC, getRobotImages, getRegionalEventDetail,
  fetchTeamRemappings, and private helpers (trimArray,
  conformFTCOfflineTeam, fetchTBATeams, awardsHilight). Includes
  epoch guard on getTeamList to prevent stale data from superseded
  requests.
- src/utils/calculateBlueBanners.js (292 lines): Pure function
  extracted from App.jsx, shared by both App.jsx and useTeamData.

Architecture:
- useTeamData is instantiated in EventStoreProvider as a 'team slice'
  that composes into the rankings slice (teamSlice feeds getEPA,
  getEPAFTC, getTeamList, getRegionalEventDetail into
  useRankingsAlliances via composedRankingsDeps).
- Team functions are exposed via both storeRef (for App.jsx callers
  above the provider tree) and EventActionsContext (for page consumers
  below the tree).
- Two narrow callbacks (requestCommunityUpdatesForTeams,
  patchSelectedEvent) replace broad setter access, keeping
  cross-concern boundaries clean.
- patchSelectedEvent uses JSON comparison for no-op safety.

Prop drilling removed (15 assignments eliminated):
- EPA, regionalEventDetail, getRegionalEventDetail moved from props
  to useEventData()/useEventActions() context in RanksPage,
  AnnouncePage, and PlayByPlayPage.
- robotImages moved from props to useEventData() in CheatsheetPage.
- playoffs moved from props to useEventData() in RanksPage.

Also removed unused inWorldChamps function from App.jsx (logic now
inlined in useTeamData).

TeamAvatar fix:
- Fixed race condition where avatars failed to render on initial page
  load. The module-level cache now stores Promises for in-flight
  fetches (instead of empty strings), and the cache is read inside
  the useEffect rather than as a dependency, preventing parent
  re-renders from cancelling pending .then() subscriptions.

App.jsx: ~5,530 -> ~4,431 lines (-1,099 lines, -20%)
Net: -1,517 lines deleted, +173 lines inserted (tracked files)
Build: zero warnings
